tests: move test model in test/utils
