+++ test/perf/append-list/add/100-items.spec.js
@@ -5,7 +5,7 @@ import { test } from 'node:test'
 import { createHistogram } from 'node:perf_hooks'
 
 import { Repository, utils } from '../../../../index.js'
-import { Chatroom } from '../../../model/index.js'
+import { Chatroom } from '../../../utils/model/index.js'
 
 test('perf: add 100 AppendList items ~ 3 kb each', async t => {
   const redis = new ioredis()
