+++ test/perf/append-list/add/10k-nested-items.spec.js
@@ -6,7 +6,7 @@ import { test } from 'node:test'
 import { createHistogram } from 'node:perf_hooks'
 
 import { Repository, utils } from '../../../../index.js'
-import { Chatroom } from '../../../model/index.js'
+import { Chatroom } from '../../../utils/model/index.js'
 
 test('perf: add 10k AppendList items ~ 3 kb, nested in 100 Lists', async t => {
