+++ test/unit/lists/lazy-list/load.spec.js
@@ -2,7 +2,7 @@ import assert from 'node:assert'
 import { test } from 'node:test'
 
 import { LazyList } from '../../../../src/list.js'
-import { Message } from '../../../model/index.js'
+import { Message } from '../../../utils/model/index.js'
 
 // Mock repository, needed when calling `list.load(repository)
 //
