+++ test/unit/map/flatten/root.spec.js
@@ -2,7 +2,7 @@ import assert from 'node:assert'
 import { test } from 'node:test'
 
 import { flatten } from '../../../../src/map.js'
-import { Chatroom } from '../../../model/index.js'
+import { Chatroom } from '../../../utils/model/index.js'
 
 test('#flatten()', async t => {
   let chatroom
