+++ test/unit/repository/fetch/fetch.spec.js
@@ -3,7 +3,7 @@ import { test } from 'node:test'
 import ioredis from 'ioredis-mock'
 
 import { Repository } from '../../../../src/repository.js'
-import { Chatroom } from '../../../model/index.js'
+import { Chatroom } from '../../../utils/model/index.js'
 
 test('Repository', async t => {
   let repo, fetched
